requirements.txt
amqp==5.1.1
ansicon==1.89.0
arabic-reshaper==2.1.3
arrow==1.2.2
asgiref==3.5.2
asn1crypto==1.5.1
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.5
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-import-export==2.8.0
django-picklefield==3.1
django-timezone-field==5.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jinxed==1.2.0
kombu==5.2.4
lxml==4.9.1
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==3.5.3
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
tinycss2==1.1.1
tk==0.1.0
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
